moved examples to common
